[DEV-667] restructured mocks directories and fixed the tests
