+++ src/payment/models/bank-tx/__mocks__/bank-tx-batch.entity.mock.ts
@@ -1,4 +1,4 @@
-import { BankTxBatch } from '../../bank-tx-batch.entity';
+import { BankTxBatch } from '../bank-tx-batch.entity';
 
 const defaultBankTxBatch: Partial<BankTxBatch> = {
   transactions: [],
