+++ src/payment/models/bank-tx/__mocks__/bank-tx.entity.mock.ts
@@ -1,4 +1,4 @@
-import { BankTx } from '../../bank-tx.entity';
+import { BankTx } from '../bank-tx.entity';
 import { createDefaultBankTxBatch } from './bank-tx-batch.entity.mock';
 
 const defaultBankTx: Partial<BankTx> = {
