+++ src/payment/models/buy-crypto/entities/__mocks__/buy-crypto.entity.mock.ts
@@ -1,6 +1,6 @@
-import { createDefaultBuy } from 'src/payment/models/buy/__tests__/mock/buy.entity.mock';
-import { AmlCheck } from '../../../enums/aml-check.enum';
-import { BuyCrypto } from '../../buy-crypto.entity';
+import { createDefaultBuy } from 'src/payment/models/buy/__mocks__/buy.entity.mock';
+import { AmlCheck } from '../../enums/aml-check.enum';
+import { BuyCrypto } from '../buy-crypto.entity';
 import { createCustomBuyCryptoBatch } from './buy-crypto-batch.entity.mock';
 
 export function createDefaultBuyCrypto(): BuyCrypto {
