+++ src/payment/models/buy-crypto/services/__tests__/buy-crypto.service.spec.ts
@@ -4,16 +4,16 @@ import { BankTxRepository } from 'src/payment/models/bank-tx/bank-tx.repository'
 import { BuyRepository } from 'src/payment/models/buy/buy.repository';
 import { BuyService } from 'src/payment/models/buy/buy.service';
 import { CryptoInputType } from 'src/payment/models/crypto-input/crypto-input.entity';
-import { createCustomCryptoInput } from 'src/payment/models/crypto-input/__tests__/mock/crypto-input.entity.mock';
+import { createCustomCryptoInput } from 'src/payment/models/crypto-input/__mocks__/crypto-input.entity.mock';
 import { CryptoRouteRepository } from 'src/payment/models/crypto-route/crypto-route.repository';
 import { CryptoRouteService } from 'src/payment/models/crypto-route/crypto-route.service';
 import { createCustomCryptoRouteHistory } from 'src/payment/models/crypto-route/dto/__tests__/mock/crypto-route-history.dto.mock';
 import { SettingService } from 'src/shared/models/setting/setting.service';
 import { TestSharedModule } from 'src/shared/test.shared.module';
 import { UserService } from 'src/user/models/user/user.service';
-import { createCustomBuyCryptoHistory } from '../../dto/__tests__/mock/buy-crypto-history.dto.mock';
+import { createCustomBuyCryptoHistory } from '../../dto/__mocks__/buy-crypto-history.dto.mock';
 import { BuyCrypto } from '../../entities/buy-crypto.entity';
-import { createCustomBuyCrypto } from '../../entities/__tests__/mock/buy-crypto.entity.mock';
+import { createCustomBuyCrypto } from '../../entities/__mocks__/buy-crypto.entity.mock';
 import { BuyCryptoRepository } from '../../repositories/buy-crypto.repository';
 import { BuyCryptoBatchService } from '../buy-crypto-batch.service';
 import { BuyCryptoDexService } from '../buy-crypto-dex.service';
