+++ src/payment/models/buy-fiat/__tests__/buy-fiat.service.spec.ts
@@ -3,14 +3,14 @@ import { Test, TestingModule } from '@nestjs/testing';
 import { TestSharedModule } from 'src/shared/test.shared.module';
 import { UserService } from 'src/user/models/user/user.service';
 import { BankTxRepository } from '../../bank-tx/bank-tx.repository';
-import { createCustomCryptoInput } from '../../crypto-input/__tests__/mock/crypto-input.entity.mock';
+import { createCustomCryptoInput } from '../../crypto-input/__mocks__/crypto-input.entity.mock';
 import { SellRepository } from '../../sell/sell.repository';
 import { SellService } from '../../sell/sell.service';
 import { BuyFiat } from '../buy-fiat.entity';
 import { BuyFiatRepository } from '../buy-fiat.repository';
 import { BuyFiatService } from '../buy-fiat.service';
 import { createCustomBuyFiatHistory } from '../dto/__tests__/mock/buy-fiat-history.dto.mock';
-import { createCustomBuyFiat } from './mock/buy-fiat.entity.mock';
+import { createCustomBuyFiat } from '../__mocks__/buy-fiat.entity.mock';
 
 enum MockBuyData {
   DEFAULT,
