+++ src/payment/models/buy/__mocks__/buy.entity.mock.ts
@@ -1,7 +1,7 @@
 import { Deposit } from 'src/payment/models/deposit/deposit.entity';
 import { createDefaultAsset } from 'src/shared/models/asset/__tests__/mock/asset.entity.mock';
-import { createDefaultUser } from 'src/user/models/user/__tests__/mock/user.entity.mock';
-import { Buy } from '../../buy.entity';
+import { createDefaultUser } from 'src/user/models/user/__mocks__/user.entity.mock';
+import { Buy } from '../buy.entity';
 
 export function createDefaultBuy(): Buy {
   return createCustomBuy({});
