+++ src/payment/models/staking/__tests__/mock/staking.entity.mock.ts
@@ -1,4 +1,4 @@
-import { createDefaultUser } from 'src/user/models/user/__tests__/mock/user.entity.mock';
+import { createDefaultUser } from 'src/user/models/user/__mocks__/user.entity.mock';
 import { Staking } from '../../staking.entity';
 
 const defaultStaking: Partial<Staking> = {
