+++ src/user/models/user-data/__tests__/mock/user-data.entity.mock.ts
@@ -1,4 +1,4 @@
-import { createDefaultCountry } from 'src/shared/models/country/__tests__/mock/country.entity.mock';
+import { createDefaultCountry } from 'src/shared/models/country/__mocks__/country.entity.mock';
 import { AccountType } from '../../account-type.enum';
 import { KycState, KycStatus, UserData } from '../../user-data.entity';
