+++ src/user/models/user/__mocks__/user.entity.mock.ts
@@ -1,6 +1,6 @@
 import { createDefaultUserData } from 'src/user/models/user-data/__tests__/mock/user-data.entity.mock';
 import { createDefaultWallet } from 'src/user/models/wallet/__tests__/mock/wallet.entity.mock';
-import { User } from '../../user.entity';
+import { User } from '../user.entity';
 
 export function createDefaultUser(): User {
   return createCustomUser({});
